requirements.txt
pillow==11.3.0
pygame
matplotlib
